+++ packages/pod/CHANGELOG.md
@@ -1,3 +1,3 @@
-## 1.0.0
+## 0.14.0
 
-- Initial version.
+- Riverpod 0.14.0 typedefs